refactor(secondary-screen): promote Now Playing panel to a widget
Extract `_buildNowPlayingPanel` (plus its `_buildScreenshotButton` and
`_buildLauncherButton` sub-builders) out of `_SecondaryScreenState` into a new
`const`-constructor `NowPlayingPanel` widget under `secondary_screen/widgets/`.

Now that the app dock is already its own widget (#139), the panel is a clean,
input-driven subtree: the host passes the state snapshot, the live session
readout (`sessionRunning` + pre-formatted `sessionTime`, both still derived from
the host-owned `_sessionWatch`), and six action callbacks (screenshot, three
dock-slot callbacks, open-launcher, open-accessibility-settings). The widget
re-reads no state of its own.

Bodies moved verbatim; the only edits are the session/callback parameter
substitutions. Host `secondary_screen.dart` shrinks ~215 lines. No behavior or
public-API change.
